i have updated all the serviceimpl in that create accounts, get, get all, update and delete. i also created controller for all the elements and add the GetMapping, PostMaping, PutMapping and DeleteMapping for all the elements. i also add ExeptionHandling in serviceimple and in also in delete section of Controller.

+++ src/main/java/com/personal/demoProject/dto/student/StudentRequestDTO.java
@@ -7,7 +7,7 @@ public class StudentRequestDTO {
     private String firstName;
     private String lastName;
     private  String address;
-    private Integer phoneNo;
+    private Long phoneNo;
     private String guardianName; // use (FK)
     private String emailId;
 }
